requirements.txt
Pillow
binance-connector